requirement.txt
streamlit
langchain
openai
pdfplumber
qdrant-client
langchain-openai
langchain-community
python-dotenv
